+++ contracts/registries/datasets/Dataset.sol
@@ -5,6 +5,10 @@ pragma solidity ^0.6.0;
 
 import "../RegistryEntry.sol";
 
+/**
+ * @dev Referenced in the SDK with the current path `contracts/registries/datasets/Dataset.sol`.
+ * Changing the name or the path would cause a breaking change in the SDK.
+ */
 contract Dataset is RegistryEntry {
     /**
      * Members