+++ contracts/registries/datasets/DatasetRegistry.sol
@@ -6,6 +6,10 @@ pragma solidity ^0.6.0;
 import "../Registry.sol";
 import "./Dataset.sol";
 
+/**
+ * @dev Referenced in the SDK with the current path `contracts/registries/datasets/DatasetRegistry.sol`.
+ * Changing the name or the path would cause a breaking change in the SDK.
+ */
 contract DatasetRegistry is Registry {
     /**
      * Constructor