+++ contracts/tools/testing/slither/Slither.sol
@@ -3,9 +3,9 @@
 
 pragma solidity ^0.8.0;
 
-import {IexecPocoAccessorsFacet} from "../../../modules/facets/IexecPocoAccessorsFacet.sol";
-import {IexecPoco1Facet} from "../../../modules/facets/IexecPoco1Facet.sol";
-import {IexecPoco2Facet} from "../../../modules/facets/IexecPoco2Facet.sol";
+import {IexecPocoAccessorsFacet} from "../../../facets/IexecPocoAccessorsFacet.sol";
+import {IexecPoco1Facet} from "../../../facets/IexecPoco1Facet.sol";
+import {IexecPoco2Facet} from "../../../facets/IexecPoco2Facet.sol";
 
 /**
  * @notice This contract is dedicated to slither analysis.