+++ CHANGELOG.md
@@ -2,14 +2,20 @@
 
 ## vNEXT
 
+- Save IexecLibOrders_v5 in config file (#242)
 - Migrate proxy to Diamond pattern (ERC-2535):
+  - Use lib as storage. (#243)
+  - Restore compatibility with iExec SDK. (#240)
+  - Target latest EVM version (#239)
+  - Adapt contracts file tree (#238)
+  - Use namespaced storage (#236, #237)
+  - Fix script folder (#235)
   - Format all solidity files (#233)
   - Rename ERC1538 architure to diamond Proxy architecture(#226, #229, #230, #234)
   - Remove ENS module (#225)
   - Add Diamond contract unit tests (#224)
   - Fix `fallback` and `receive` (#223)
   - Migrate contracts (#222)
-
 - Add Github Action CI in order to publish NPM package
 
 ### Updated contracts